feat(page): Allow a custom preview selector
Allow a custom preview selector which will select a different element on
the page to power the preview. This is useful in scenarios where the
markup is outside the example component's markup (e.g. a lightbox, or
modal). If the selector is not present, it will fallback to using the
preview element (by using `[data-preview]`).

It can be used by providing a css selector in the js props:
`"data-preview-selector": ".foo-class"`.